Refactor into smaller files
